// src/api/openapi/model/offerStatementEntryOutput.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferStatementEntryEntry } from './offerStatementEntryEntry';


export interface OfferStatementEntryOutput { 
    children?: Array<OfferStatementEntryOutput>;
    entries?: Array<OfferStatementEntryEntry>;
    id: string;
    name: string;
    notMade: boolean;
    originalPrice: number;
    price: number;
}